[WebDriverAgent] Cache current alert when checking if elements are obstructed
Summary:
`_currentAlert` is pretty slow because it searches the hierarchy (by pretty slow we mean like 2 seconds, mainly because of UIAElement stuff). This search method is "linear" with respect to number of elements if finds except that it is more like quadratic because `_currentAlert` gets slower as there are more elements in the tree, and you expect that more results are found as there are more elements.

Anyways, with Xcode 7 this just takes way too long to finish.