Faster alert checks for iPhones
Summary:
In case of iPhone any alert or sheet is considered an alert.
In case of iPad any alert and sheet not nested in popover is considered an alert. This is because there is legit use case for sheets in popover. I think in the future we might want to stop treating sheet as alerts at all.

When running WDA on iPhone we do NOT need keep quering whether sheets are contained by popover as they do not exist on iPhones.
We can save some time if we check for device first and than query popover case.

Reviewed By: lawrencelomax

Differential Revision: D3620860

fbshipit-source-id: 793b5bb415c07f2f077b778891ef929fee0ebed1