Change iPad used for integration tests to iPad 2
Summary:
`-[FBKeyboardTests testTextTyping]` is failing on slower hosts, because text field is cleared in the middle of typing which finally causes string mismatch, when checking entered text.
This is known reported XCTest issue. Only workaround is to use simulator with smaller screen, which performs faster. In that case iPad 2

Closes https://github.com/facebook/WebDriverAgent/pull/233

Reviewed By: lawrencelomax

Differential Revision: D3669325

Pulled By: marekcirkos

fbshipit-source-id: 2bc0b940d452a0cee3191de3a32249706c4d7593