Send string to picker when setting value
Summary:
I'm not at all certain that this is the correct way to do this, but it fixes the problem.

When sending text to a PickerWheel, in order to set it, the array value is not found in the possible values and it is rejected. So concatenate the array into a string.

See https://github.com/appium/appium/issues/6873
Closes https://github.com/facebook/WebDriverAgent/pull/295

Differential Revision: D3907080

Pulled By: marekcirkos

fbshipit-source-id: 11bb2201e3672b8ad5c3eb826fcb6a573aa2ef6a