More robust testDeactivateApplication
Summary:
Visibility is a bit "delayed". Even if we wait till different view on that screen is visible, test can still fail.
So I think that checking for element existance right after `fb_deactivateWithDuration` and that waiting till element is visible should be sufficient for checking if that command worked.

Reviewed By: migchez

Differential Revision: D3923189

fbshipit-source-id: 4d3956f9d94b7d8b69f671b73f81fd51f9519440