Improve performance of visibleCells endpoint
Summary:
Number of returned cells is potentialy very big, which means that in order to prepare response we will need to resolve every single element.
Resolving element is not cheap. So in order to make it faster we can create response that returns only cached element ids, without resolving them.

Reviewed By: migchez

Differential Revision: D3950992

fbshipit-source-id: e9515074fb73edc305ae3c01d271909cc145aaa1